♻️ Re-organize CI and package
