+++ .github/workflows/build.yml
@@ -25,7 +25,7 @@ jobs:
       - id: set-matrix
         shell: bash
         run: |
-          BASE_GROUPS=$(jq -n -c '["core"]')
+          BASE_GROUPS=$(jq -n -c '["core", "agents"]')
           # Run modal on push or when modal paths change
           if [[ "${{ github.event_name }}" == "push" || "${{ steps.changes.outputs.modal }}" == "true" ]]; then
             MATRIX=$(jq -n -c --argjson groups "$BASE_GROUPS" '{group: ($groups + ["modal"])}')
@@ -73,6 +73,8 @@ jobs:
           LAMIN_API_KEY: ${{ secrets.LAMIN_API_KEY_TESTUSER1 }}
       - run: pytest tests/core
         if: ${{ matrix.group == 'core' }}
+      - run: pytest tests/agents
+        if: ${{ matrix.group == 'agents' }}
       - run: pytest tests/modal
         if: ${{ matrix.group == 'modal' }}
